reorganize codebase for admin widgets
<!-- ps-id: 1f6c4534-6274-43b9-8b6c-27178e326ffc -->
