+++ apps/admin-api/src/users/users.module.ts
@@ -1,11 +1,9 @@
 import { Module } from '@nestjs/common';
 import { UsersService } from '@/users/users.service';
 import { PrismaModule } from '@/prisma/prisma.module';
-import { UsersController } from '@/users/users.controller';
 
 @Module({
   imports: [PrismaModule],
-  controllers: [UsersController],
   providers: [UsersService],
   exports: [UsersService],
 })
