+++ apps/admin-api/src/users/users.service.ts
@@ -1,7 +1,6 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { User } from '@prisma/client';
 import { PrismaService } from '@/prisma/prisma.service';
-import * as bcrypt from 'bcryptjs';
 
 @Injectable()
 export class UsersService {
@@ -37,26 +36,5 @@ export class UsersService {
         refreshToken,
       },
     });
-  }
-
-  async changePassword(userId: number, password: string) {
-    const user = await this.prisma.user.findUnique({
-      where: { id: userId },
-    });
-
-    if (!user) {
-      throw new BadRequestException('User not found');
-    }
-
-    // Hash the new password
-    const hashedPassword = bcrypt.hashSync(password, 10);
-
-    // Update the user's password in the database
-    await this.prisma.user.update({
-      where: { id: userId },
-      data: { password: hashedPassword },
-    });
-
-    return { message: 'Password changed successfully' };
   }
 }
