+++ apps/admin-api/src/widgets/users/users-widget.service.ts
@@ -1,0 +1,29 @@
+import { BadRequestException, Injectable } from '@nestjs/common';
+import { PrismaService } from '@/prisma/prisma.service';
+import * as bcrypt from 'bcryptjs';
+
+@Injectable()
+export class UsersWidgetService {
+  constructor(private readonly prisma: PrismaService) {}
+
+  async changePassword(userId: number, password: string) {
+    const user = await this.prisma.user.findUnique({
+      where: { id: userId },
+    });
+
+    if (!user) {
+      throw new BadRequestException('User not found');
+    }
+
+    // Hash the new password
+    const hashedPassword = bcrypt.hashSync(password, 10);
+
+    // Update the user's password in the database
+    await this.prisma.user.update({
+      where: { id: userId },
+      data: { password: hashedPassword },
+    });
+
+    return { message: 'Password changed successfully' };
+  }
+}
