+++ apps/admin-client/src/components/fields/integer-field.tsx
@@ -1,5 +1,5 @@
 import { ChangeEvent } from 'react';
-import FieldLabel from '../../features/records/record-field-label';
+import FieldLabel from './record-field-label';
 import { AdminModelField } from '@repo/types';
 import { Input } from '@/components/input';
