+++ apps/admin-client/src/components/fields/select-field.tsx
@@ -1,4 +1,4 @@
-import FieldLabel from '../../features/records/record-field-label';
+import FieldLabel from './record-field-label';
 import { AdminModelField, SelectOption } from '@repo/types';
 import {
   Select,
