+++ apps/admin-client/src/features/records/show/delete-record-card.tsx
