+++ apps/admin-client/src/features/records/show/show-view-field.tsx
